Prebuild list of redundancies
